# internal/modules/webhooks/handler.go
package webhooks

import (
	"encoding/json"
	"net/http"
	"time"

	"github.com/deveasyclick/iwifunni/internal/shared/validate"
	"github.com/deveasyclick/iwifunni/internal/shared/authctx"
	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
)

type Handler struct {
	service *Service
}

func NewHandler(service *Service) *Handler {
	return &Handler{service: service}
}

func (h *Handler) Register(r chi.Router) {
	r.Post("/webhooks", h.create)
	r.Get("/webhooks", h.list)
	r.Delete("/webhooks/{webhookID}", h.delete)
}

type createRequest struct {
	URL    string   `json:"url" validate:"required,url"`
	Events []string `json:"events" validate:"required,min=1"`
	Secret string   `json:"secret" validate:"required"`
}

type webhookResponse struct {
	ID        uuid.UUID `json:"id"`
	URL       string    `json:"url"`
	Events    []string  `json:"events"`
	IsActive  bool      `json:"is_active"`
	CreatedAt time.Time `json:"created_at"`
}

// @Summary      Create webhook
// @Description  Register a new webhook endpoint for notification events
// @Tags         Webhooks
// @Accept       json
// @Produce      json
// @Param        body  body  createRequest  true  "Webhook URL and events"
// @Success      201   {object}  webhookResponse
// @Failure      400   {string}  string  "Invalid input"
// @Failure      401   {string}  string  "Unauthorized"
// @Router       /webhooks [post]
// @Security     BearerAuth
func (h *Handler) create(w http.ResponseWriter, r *http.Request) {
	environmentID, ok := authctx.GetEnvironmentID(r.Context())
	if !ok {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}
	var req createRequest
	if !validate.DecodeAndRespond(w, r, &req) {
		return
	}
	wh, err := h.service.Create(r.Context(), CreateInput{
		EnvironmentID: environmentID,
		URL:           req.URL,
		Secret:        req.Secret,
		Events:        req.Events,
	})
	if err != nil {
		http.Error(w, "failed to register webhook", http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	_ = json.NewEncoder(w).Encode(webhookResponse{ID: wh.ID, URL: wh.Url, Events: wh.Events, IsActive: wh.IsActive, CreatedAt: wh.CreatedAt.Time})
}

// @Summary      List webhooks
// @Description  Get all registered webhooks
// @Tags         Webhooks
// @Produce      json
// @Success      200  {array}   webhookResponse
// @Failure      401  {string}  string  "Unauthorized"
// @Router       /webhooks [get]
// @Security     BearerAuth
func (h *Handler) list(w http.ResponseWriter, r *http.Request) {
	environmentID, ok := authctx.GetEnvironmentID(r.Context())
	if !ok {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}
	whs, err := h.service.List(r.Context(), environmentID)
	if err != nil {
		http.Error(w, "failed to list webhooks", http.StatusInternalServerError)
		return
	}
	resp := make([]webhookResponse, 0, len(whs))
	for _, wh := range whs {
		resp = append(resp, webhookResponse{ID: wh.ID, URL: wh.Url, Events: wh.Events, IsActive: wh.IsActive, CreatedAt: wh.CreatedAt.Time})
	}
	w.Header().Set("Content-Type", "application/json")
	_ = json.NewEncoder(w).Encode(resp)
}

// @Summary      Delete webhook
// @Description  Unregister a webhook
// @Tags         Webhooks
// @Param        webhookID  path  string  true  "Webhook ID"
// @Success      204        {string}  string  "No content"
// @Failure      400        {string}  string  "Invalid ID"
// @Failure      401        {string}  string  "Unauthorized"
// @Router       /webhooks/{webhookID} [delete]
// @Security     BearerAuth
func (h *Handler) delete(w http.ResponseWriter, r *http.Request) {
	environmentID, ok := authctx.GetEnvironmentID(r.Context())
	if !ok {
		http.Error(w, "unauthorized", http.StatusUnauthorized)
		return
	}
	webhookID, err := uuid.Parse(chi.URLParam(r, "webhookID"))
	if err != nil {
		http.Error(w, "invalid webhook id", http.StatusBadRequest)
		return
	}
	if err := h.service.Delete(r.Context(), webhookID, environmentID); err != nil {
		http.Error(w, "failed to delete webhook", http.StatusInternalServerError)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}